ItemsControl - don't reapply the same template
Certain Styles application sequences can lead to the same ItemsPanelTemplate (but not reference-equal) being applied multiple times.

As a perf improvement, don't reapply the panel. This also works around the issue on Android that changing the ItemsPanel of a ListView after it's already been materialized isn't properly supported and causes errors.